+++ lib/ui/sea_creature/sea_creature_screen.dart
@@ -103,6 +103,9 @@ class SeaCreatureScreen extends BaseScreen<SeaCreaturesProvider, SeaCreature> {
   @override
   List<InfoChip> listInformationWidgets(BuildContext context, SeaCreature item) {
     final calendarOptions = Provider.of<PreferencesProvider>(context, listen: false).calendarOptions;
+    final isCurrentlyAvailable = item.isCurrentlyAvailable(calendarOptions);
+    final isNewlyAvailable = item.isNewlyAvailable(calendarOptions);
+    final isLeavingSoon = item.isLeavingSoon(calendarOptions);
 
     return [
       InfoChip(
@@ -118,7 +121,17 @@ class SeaCreatureScreen extends BaseScreen<SeaCreaturesProvider, SeaCreature> {
       InfoChip(
         title: "Available",
         color: AppColors.currentlyAvailableColor,
-        visible: item.isCurrentlyAvailable(calendarOptions),
+        visible: isCurrentlyAvailable && !(isNewlyAvailable || isLeavingSoon),
+      ),
+      InfoChip(
+        title: "New",
+        color: AppColors.newlyAvailableColor,
+        visible: isNewlyAvailable,
+      ),
+      InfoChip(
+        title: "Leaving soon",
+        color: AppColors.leavingSoonColor,
+        visible: isLeavingSoon,
       ),
     ];
   }
@@ -126,6 +139,9 @@ class SeaCreatureScreen extends BaseScreen<SeaCreaturesProvider, SeaCreature> {
   @override
   List<InfoBadge> gridInformationWidgets(BuildContext context, SeaCreature item) {
     final calendarOptions = Provider.of<PreferencesProvider>(context, listen: false).calendarOptions;
+    final isCurrentlyAvailable = item.isCurrentlyAvailable(calendarOptions);
+    final isNewlyAvailable = item.isNewlyAvailable(calendarOptions);
+    final isLeavingSoon = item.isLeavingSoon(calendarOptions);
 
     return [
       InfoBadge(
@@ -140,7 +156,17 @@ class SeaCreatureScreen extends BaseScreen<SeaCreaturesProvider, SeaCreature> {
       ),
       InfoBadge(
         color: AppColors.currentlyAvailableColor,
-        visible: item.isCurrentlyAvailable(calendarOptions),
+        visible: isCurrentlyAvailable && !(isNewlyAvailable || isLeavingSoon),
+        alignment: Alignment.topLeft,
+      ),
+      InfoBadge(
+        color: AppColors.newlyAvailableColor,
+        visible: isNewlyAvailable,
+        alignment: Alignment.topLeft,
+      ),
+      InfoBadge(
+        color: AppColors.leavingSoonColor,
+        visible: isLeavingSoon,
         alignment: Alignment.topLeft,
       ),
     ];